Fluent validation updates (#89)
* changed lat/long validators and tests to use decimal instead of double
* simplified validators

+++ src/JK.Common.FluentValidation.Tests/MockModel.cs
@@ -5,6 +5,6 @@ namespace JK.Common.FluentValidation.Tests;
 public class MockModel
 {
     public DateTime DateValue { get; set; }
-    public double DoubleValue { get; set; }
+    public decimal DecimalValue { get; set; }
     public string StringValue { get; set; }
 }
